move MethodMembersVisitor to classes module
