+++ tests/test_visitors/test_ast/test_complexity/test_classes/test_method_counts.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from wemake_python_styleguide.visitors.ast.complexity.counts import (
+from wemake_python_styleguide.visitors.ast.complexity.classes import (
     MethodMembersVisitor,
     TooManyMethodsViolation,
 )
